requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.4
django-environ==0.8.1
django-tinymce==3.4.0
mysqlclient==2.1.0
Pillow==9.1.0
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
